requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
gunicorn==20.1.0
dj-database-url>=1.0,<2.0
whitenoise